requirements.txt
aiohttp==3.8.4
python-dotenv==1.1.1
